Changed: move dashboard on a top level
