+++ src/app/dashboard/dashboard.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'xp-dashboard',
+    template: `
+        dashboard works.
+    `
+})
+export class DashboardComponent implements OnInit {
+
+    constructor() { }
+
+    ngOnInit() {
+    }
+
+}
